requirements.txt
asgiref==3.8.1
Django==5.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-rest-framework==0.1.0
django-resumable==0.2.0
djangorestframework==3.15.2
pillow==10.4.0
python-magic==0.4.27
sqlparse==0.5.1
typing_extensions==4.12.2
